requirements.txt
streamlit==1.29.0
google-generativeai==0.3.1
python-dotenv==1.0.0
validators==0.22.0
beautifulsoup4==4.12.2
nltk==3.8.1
requests==2.31.0
lxml-html-clean
trafilatura==1.6.1